pyproject.toml
[tool.poetry]
name = "posttelegram"
version = "0.1.0"
description = ""
authors = ["ren <denis30510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
telebot = "^0.0.5"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
ipython = "^8.23.0"
djangorestframework = "^3.15.1"
django-celery-beat = "^2.6.0"
celery = "^5.4.0"
redis = "^5.0.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
